Refactor: Separate each action into its own module
Previously, all actions executable by the reducer were implemented
within the `reducer.py` module. This made the file long, hard to
maintain, and obscured the core logic behind the implementation
details of the actions.

With this commit, each action now resides in its own module. The
`reducer.py` file has been redesigned to be as independent as
possible from the specific implementations of the actions.
As a result, adding a new action or removing an existing one now
requires modifying only two lines in `reducer.py`. Furthermore,
changes to an individual action will likely not require any
changes to `reducer.py` at all.

To support this modular structure, each action module must follow
a specific protocol to be compatible with the reducer.
Specifically, each module must:

- Define a `configure_subparser` function to set up its
  command-line subparser.
- Define an `ACTION_NAME` variable specifying the name of the
  action it implements.
- Implement a function with the same name as `ACTION_NAME`, which
  will be the entry point called by the reducer.

Another notable change introduced by this commit is the enforcement
of using the `medunda.` prefix when importing modules. This is a
necessary step toward making the project distributable as a package.
However, it comes with a minor inconvenience: the code must now be
executed from the `src` directory using the following command:

```python
python3 -m medunda.reducer ...
```