Introducing the concept of Dataset
This commit adds support for the concept of a `Dataset`,
which is a collection of files representing the same domain
across different variables or time points.

A `Dataset` has a root directory that contains all its related
files. Within this directory, a JSON file defines the structure
and metadata of the `Dataset`.

The creation of a `Dataset` is handled by the `downloader`. Before
starting the actual data download, the `downloader` generates the
JSON file. This design enables the ability to resume interrupted
downloads or re-download corrupted files.