Ensuring that the argument of computing_layer_height is a Numpy array
We declared that the argument of the computing_layer_height
function is an "ArrayLike" object. This includes Numpy arrays, but
also tuples and lists.
In the function, instead, we take for granted that the argument
is an array. We use "asarray" to transform the object into an
array when it is a list of a tuple.